bayesbybackprop/requirements.txt
tensorflow-gpu==1.7.0
